Prestige Roshanara Whitefield Bangalore
Prestige Roshanara Whitefield Bangalore is a high end unrivaled residential development located in South Bangalore, one of the most exclusive residential areas of the city. A place which is most famous as a commercial as well as residential hotspot in Bangalore.

Whitefield in South Bangalore, a residential landmark is now available for pre-launch bookings at attractive pre-launch prices. It is a luxury condominium project and best example of modern architectural design with traditional aspects.

Best residential projects in Bangalore offer a new way of life and a comfortable life for every age group. Step into the dream space for a thriving future.

The property in Bangalore ensures that it will be the ideal location for those who want to enjoy sophisticated modern living without sacrificing their traditions. It is conveniently located near the airport and has excellent connectivity. These accommodations are only for those who value luxury and quality. It is just 5 minutes drive away from metro stations. One can easily take advantage of proximity to important places.

Prestige Roshanara Whitefield Floor plan It offers 1BHK, 2BHK, and 3BHK residential flats, each with a distinct ambiance. The complex is one acre wide in size and has a large number of apartments. It is one of the most popular suburbs of the city. These exquisite apartments are well built with muscular and modern structures, and all units include Vastu compliant smart homes. This project affects Vastu dwellings in all units.

Prestige Roshanara review is fantastic as it is a perfect home for home buyers. The luxury condominium project best exemplifies modern architectural design with traditional aspects, ensuring that it will be the ideal venue for those who like to live a sophisticated modern life without surrendering the conventions that have represented their identity up to this point.

For More Details
Visit us - https://bit.ly/3BCPbxe